// src/api/models/MfaBeginRequestBody.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DeviceInfo } from './DeviceInfo';
import type { User } from './User';
export type MfaBeginRequestBody = {
    deviceInfo?: DeviceInfo;
    trustInfo: string;
    user: User;
};